Tabs: Scope the tab movers to the Tab List block
The mover buttons only make sense where tabs are shown and selected, so move
them out of the shared TabToolbarControls (rendered by Tabs, Tab List, and Tab
Panels) into a dedicated TabMovers component rendered solely by the Tab List
block. TabToolbarControls goes back to only the Add/Remove buttons.

Relocate the mover stylesheet and its class to the Tab List block as well
(wp-block-tab-list__mover-button, registered via the tab-list editorStyle) so
the styling lives with the block that owns the buttons.
