+++ packages/block-library/src/editor.scss
@@ -48,8 +48,8 @@
 @use "./social-link/editor.scss" as *;
 @use "./social-links/editor.scss" as *;
 @use "./spacer/editor.scss" as *;
+@use "./tab-list/editor.scss" as *;
 @use "./table/editor.scss" as *;
-@use "./tabs/editor.scss" as *;
 @use "./template-part/editor.scss" as *;
 @use "./term-template/editor.scss" as *;
 @use "./text-columns/editor.scss" as *;
