+++ packages/block-library/src/tab-list/block.json
@@ -77,5 +77,6 @@
 			"padding": ".wp-block-tab-list button"
 		}
 	},
-	"style": "wp-block-tab-list"
+	"style": "wp-block-tab-list",
+	"editorStyle": "wp-block-tab-list-editor"
 }
