+++ packages/block-library/src/tabs/block.json
@@ -45,6 +45,5 @@
 	},
 	"usesContext": [ "core/tabs-list", "core/tabs-id" ],
 	"style": "wp-block-tabs",
-	"editorStyle": "wp-block-tabs-editor",
 	"viewScriptModule": "@wordpress/block-library/tabs/view"
 }
